add external opener
