+++ WORKLOG.md
@@ -2,6 +2,10 @@
 
 ## 2026-04-28
 
+### nx2 canvas — library vs extension panel split
+- **`nx-panel-library.js`**: OOTB block library / templates / icons / placeholders UI (fetch, insert, preview, sprites); shares **`nx-panel-extensions.css`** with the iframe host.
+- **`nx-panel-extensions.js`**: **`nx-panel-extension`** only chooses **`nx-panel-library`** vs BYO **`iframe`** + **`iframe-protocol`**.
+
 ### nx2 canvas — tool panel sections (Editor / Library / Extensions)
 - **`helpers.js`**: **`getCanvasToolPanelViews`** — Editor placeholder tab (`editor-coming-soon`), **Library** = OOTB plugins + **`aem-assets`** (sorted **`blocks` → `aem-assets` → `icons` → `templates` → `placeholders`**), **Extensions** = other configured plugins.
 - **`tool-panel.js` / `.css`**: Picker items built with **`nx-picker`** **`section`** headings; initial tab is **`views[0]`**; prune **`_loaded`** / clear content when **`views`** empty or ids change. Placeholder host class **`.nx-tool-panel-editor-placeholder`**.
@@ -19,6 +23,14 @@
 ### nx2 canvas — AEM Assets Cancel closes panel
 - **`aem-assets.js`**: pass **`onClose`** through to **`PureJSSelectors.renderAssetSelector`** (same hook as da.live **`da-assets.js`**).
 - **`nx-panel-extensions.js`**: **`onClose`** dispatches **`nx-panel-close`** so **`panel.js`** hides the right aside.
+
+### nx2 canvas — `experience` for picker / tab bypass (`window`, `fullsize-dialog`)
+- **`helpers.js`**: **`extensionToPanelView`** passes through **`experience`** and **`sources`** from the extension config (no separate URL / modal flags).
+- **`aem-assets.js`**: **`getAssetsPlugin`** uses **`experience: 'fullsize-dialog'`** (was **`aem-assets`**).
+- **`picker.js` / `.css`**: **`experience === 'window'`** + **`sources[0]`** → new tab; **`fullsize-dialog`** → **`nx-picker-experience-dialog`** (no **`change`**); open-in icon for those rows.
+- **`tool-panel.js` / `.css`**: same rules in **`_activate`** / **`showView`**; **`_fullsizeDialogViewId`** drives **`.tool-panel-fullsize-dialog`**; body mounts **`await view.load()`**. **`@nx-panel-close`** on **`dialog`** stops propagation and closes the dialog (not the whole panel).
+- **`nx-panel-extensions.js`**: **`fullsize-dialog` + `aem-assets`** renders the assets host div and runs **`renderAssets`** from **`updated`**; other **`fullsize-dialog`** third-party configs use the iframe path as today.
+- **`nx-panel-extensions.js`**: no inline AEM Assets mount (modal-only).
 
 ## 2026-04-27
