+++ nx2/blocks/tool-panel/tool-panel.css
@@ -68,3 +68,66 @@
 .nx-tool-panel-editor-placeholder {
   min-height: 120px;
 }
+
+/* Full-size experience dialog (same shell as nx-panel-extension preview). */
+.tool-panel-fullsize-dialog {
+  position: fixed;
+  inset: 0;
+  width: 90vw;
+  max-width: 1200px;
+  height: 80vh;
+  margin: auto;
+  padding: 0;
+  border: 1px solid light-dark(var(--s2-gray-300), var(--s2-gray-600));
+  border-radius: 8px;
+  background: light-dark(var(--s2-gray-50), var(--s2-gray-800));
+  color: inherit;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.tool-panel-fullsize-dialog::backdrop {
+  background: rgb(0 0 0 / 40%);
+}
+
+.tool-panel-fullsize-dialog-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 12px 16px;
+  border-bottom: 1px solid light-dark(var(--s2-gray-200), var(--s2-gray-600));
+  flex-shrink: 0;
+}
+
+.tool-panel-fullsize-dialog-title {
+  margin: 0;
+  font-size: 14px;
+  font-weight: 600;
+}
+
+.tool-panel-fullsize-dialog-close {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 28px;
+  height: 28px;
+  padding: 0;
+  border: none;
+  border-radius: 4px;
+  background: none;
+  cursor: pointer;
+  font-size: 16px;
+  color: inherit;
+}
+
+.tool-panel-fullsize-dialog-close:hover {
+  background: light-dark(var(--s2-gray-200), var(--s2-gray-600));
+}
+
+.tool-panel-fullsize-dialog-body {
+  flex: 1;
+  position: relative;
+  min-height: 0;
+  overflow: hidden;
+}
